Add documentation comments to all public classes and methods in Debugger assembly that didn't already have them, and are not self-explanatory.
Separate various types into their own source files for clarity.

Remove DebugValueRepresentationKind.

Rename DebugBreakpointLocation to DebugSourceLocation.

Refactor DebugValueEvaluationResultKind to have names that actually describe its meaning and purpose.
