+++ src/Debugger/Engine/Impl/AD7PendingBreakpoint.cs
@@ -57,7 +57,7 @@ namespace Microsoft.R.Debugger.Engine {
             TEXT_POSITION start, end;
             GetLocation(out fileName, out lineNumber, out start, out end);
 
-            _boundBreakpoint = new AD7BoundBreakpoint(this, new DebugBreakpointLocation(fileName, lineNumber), _state);
+            _boundBreakpoint = new AD7BoundBreakpoint(this, new DebugSourceLocation(fileName, lineNumber), _state);
             return VSConstants.S_OK;
         }
