+++ src/Debugger/Impl/DebugValueEvaluationResult.cs
@@ -14,34 +14,72 @@ using Newtonsoft.Json.Linq;
 using static System.FormattableString;
 
 namespace Microsoft.R.Debugger {
-    public enum DebugValueEvaluationResultKind {
-        UnnamedItem,
-        NamedItem,
-        Slot,
-    }
-
-    [Flags]
-    public enum DebugValueEvaluationResultFlags {
-        None,
-        Atomic = 1 << 1,
-        Recursive = 1 << 2,
-        HasParentEnvironment = 1 << 3,
-    }
-
+    /// <summary>
+    /// Describes the result of evaluating an expression that produced a value that is not a promise or an active binding. 
+    /// </summary>
+    /// <remarks>
+    /// Note that most properties of the object will only have a meaningful value if the corresponding <see cref="DebugEvaluationResultFields"/>
+    /// flag was specified when producing the result. All properties which were not so requested will be <see langword="null"/>.
+    /// </remarks>
     public class DebugValueEvaluationResult : DebugEvaluationResult {
-        public DebugValueEvaluationResultKind Kind { get; }
+        private JObject _reprObj;
+
+        /// <summary>
+        /// The kind of accessor that was used to obtain this <see cref="DebugValueEvaluationResult"/> from its parent.
+        /// </summary>
+        public DebugChildAccessorKind AccessorKind { get; }
+        /// <summary>
+        /// Type of the value, as computed by <c>typeof(...)</c>.
+        /// </summary>
         public string TypeName { get; }
+        /// <summary>
+        /// List of classes of the value, as computed by <c>classes(...)</c>.
+        /// </summary>
+        /// <remarks>
+        /// If <see cref="DebugEvaluationResultFields.Classes"/> was not specified, this property will be
+        /// <see langword="null"/>, rather than an empty collection.
+        /// </remarks>
         public IReadOnlyList<string> Classes { get; }
+        /// <summary>
+        /// Length of the value, as computed by <c>length(...)</c>.
+        /// </summary>
         public int? Length { get; }
+        /// <summary>
+        /// Number of attributes that this value has, as computed by <c>length(attributes(...))</c>.
+        /// </summary>
         public int? AttributeCount { get; }
+        /// <summary>
+        /// Number of slots that this value has, as computed by <c>length(slotNames(class(...)))</c>.
+        /// </summary>
         public int? SlotCount { get; }
+        /// <summary>
+        /// Number of names that the children of value have, as computed by <c>length(names(...))</c>.
+        /// </summary>
         public int? NameCount { get; }
+        /// <summary>
+        /// Dimensions that this value has, as computed by <c>dim(...)</c>.
+        /// </summary>
+        /// <remarks>
+        /// If <see cref="DebugEvaluationResultFields.Dim"/> was not specified, this property will be
+        /// <see langword="null"/>, rather than an empty collection.
+        /// </remarks>
         public IReadOnlyList<int> Dim { get; }
+        /// <summary>
+        /// Various miscellaneous flags describing this value.
+        /// </summary>
         public DebugValueEvaluationResultFlags Flags { get; }
 
+        /// <seealso cref="DebugValueEvaluationResultFlags.Atomic"/>
         public bool IsAtomic => Flags.HasFlag(DebugValueEvaluationResultFlags.Atomic);
+
+        /// <seealso cref="DebugValueEvaluationResultFlags.Recursive"/>
         public bool IsRecursive => Flags.HasFlag(DebugValueEvaluationResultFlags.Recursive);
+
+        /// <summary>
+        /// Whether this value has any attributes.
+        /// </summary>
         public bool HasAttributes => AttributeCount != null && AttributeCount != 0;
+
         public bool HasSlots => SlotCount != null && SlotCount != 0;
 
         /// <summary>
@@ -76,9 +114,6 @@ namespace Microsoft.R.Debugger {
             }
         }
 
-
-        private JObject _reprObj;
-
         internal DebugValueEvaluationResult(DebugSession session, string environmentExpression, string expression, string name, JObject json)
             : base(session, environmentExpression, expression, name) {
 
@@ -109,14 +144,16 @@ namespace Microsoft.R.Debugger {
             var kind = json.Value<string>("kind");
             switch (kind) {
                 case null:
+                    AccessorKind = DebugChildAccessorKind.None;
+                    break;
                 case "[[":
-                    Kind = DebugValueEvaluationResultKind.UnnamedItem;
+                    AccessorKind = DebugChildAccessorKind.Brackets;
                     break;
                 case "$":
-                    Kind = DebugValueEvaluationResultKind.NamedItem;
+                    AccessorKind = DebugChildAccessorKind.Dollar;
                     break;
                 case "@":
-                    Kind = DebugValueEvaluationResultKind.Slot;
+                    AccessorKind = DebugChildAccessorKind.At;
                     break;
                 default:
                     throw new InvalidDataException(Invariant($"Invalid kind '{kind}' in:\n\n{json}"));
@@ -142,17 +179,36 @@ namespace Microsoft.R.Debugger {
             }
         }
 
+        /// <summary>
+        /// Returns a structure containing various representations of this value, as requested via <see cref="DebugEvaluationResultFields"/>
+        /// when this result was produced.
+        /// </summary>
         public DebugValueRepresentation GetRepresentation() {
-            return GetRepresentation(DebugValueRepresentationKind.Normal);
-        }
-
-        public DebugValueRepresentation GetRepresentation(DebugValueRepresentationKind kind) {
             if (_reprObj == null) {
                 throw new InvalidOperationException("Evaluation result does not have an associated representation.");
             }
-            return new DebugValueRepresentation(_reprObj, kind);
+            return new DebugValueRepresentation(_reprObj);
         }
 
+        /// <summary>
+        /// Computes the children of this value, and returns a collection of evaluation results describing each child.
+        /// See <see cref="DebugSession.EvaluateAsync"/> for the meaning of parameters.
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// Where order matters (e.g. for children of atomic vectors and lists), children are returned in that order.
+        /// Otherwise, the order is undefined. If an object has both ordered and unordered children (e.g. it is a vector
+        /// with slots), then it is guaranteed that each group is reported as a contiguous sequence within the returned
+        /// collection, and order is honored within each group; but groups themselves are not ordered relative to each other.
+        /// </para>
+        /// <para>
+        /// This method does not respect snapshot semantics - that is, it will re-evaluate the expression that produced
+        /// its value, and will obtain the most current list of children, rather than the ones that were there when
+        /// the result was originally produced. If the context in which this result was produced is no longer available
+        /// (e.g. if it was a stack frame that has since went away), the results are undefined.
+        /// </para>
+        /// </remarks>
+        /// <exception cref="RException">Raised if child retrieval fails.</exception>
         public async Task<IReadOnlyList<DebugEvaluationResult>> GetChildrenAsync(
             DebugEvaluationResultFields fields,
             int? maxLength = null,
