--- src/Debugger/Impl/DebugValueRepresentationKind.cs
@@ -1,8 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License. See LICENSE in the project root for license information.
-
-namespace Microsoft.R.Debugger {
-    public enum DebugValueRepresentationKind {
-        Normal = 0,
-    }
-}
