+++ src/Debugger/Impl/IDebugSessionProvider.cs
@@ -6,7 +6,13 @@ using System.Threading.Tasks;
 using Microsoft.R.Host.Client;
 
 namespace Microsoft.R.Debugger {
+    /// <summary>
+    /// A service that manages debug session instances for R sessions.
+    /// </summary>
     public interface IDebugSessionProvider {
+        /// <summary>
+        /// Returns the debug session associated with the given R session. Creates the session if it doesn't exist.
+        /// </summary>
         Task<DebugSession> GetDebugSessionAsync(IRSession session, CancellationToken cancellationToken = default(CancellationToken));
     }
 }
